--- a/arbor/cable_cell.cpp
+++ b/arbor/cable_cell.cpp
@@ -169,14 +169,13 @@ struct cable_cell_impl {
     }
 
     void paint(const region& reg, const scaled_mechanism<density>& prop) {
-        mextent cables = thingify(reg, provider);
-        auto& mm = get_region_map(prop.t_mech);
-
         std::unordered_map<std::string, iexpr_ptr> im;
         for (const auto& [fst, snd]: prop.scale_expr) {
             im.insert_or_assign(fst, thingify(snd, provider));
         }
 
+        auto& mm = get_region_map(prop.t_mech);
+        const auto& cables = thingify(reg, provider);
         for (const auto& c: cables) {
             // Skip zero-length cables in extent:
             if (c.prox_pos == c.dist_pos) continue;