--- a/arbor/communication/communicator.cpp
+++ b/arbor/communication/communicator.cpp
@@ -1,4 +1,3 @@
-#include <iostream>
 #include <utility>
 #include <vector>