--- a/example/busyring/readme.md
+++ b/example/busyring/readme.md
@@ -1,0 +1,3 @@
+# Ring Example
+
+A miniapp that demonstrates how to describe how to build a simple ring network.