@boeschf 2023-03-15
Thanks for this! A couple remarks:
- Could you group the profiler timings a bit differently by grouping everything that is not "init" under a common tag (maybe "run" or the like)?
- Can you test what happens if the cell_groups are larger than 1 cell (i.e. the default when using the GPU?)