@thorstenhater 2023-03-15
2. Is already done: I did not observe any difference (using up to 8 threads) when setting group size = 1 (default) and = 10000 (that in the benchmark above `-G`)