@thorstenhater 2023-03-15
1. Is a reasonable request, but requires re-writing all the profiler tags. I'd rather make a new PR fixing the problem's root cause: We cannot have nested regions. That OK for you?